+++ src/Exception/InvalidPasswordException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace HeimrichHannot\LoginRegistrationBundle\Exception;
+
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
+
+class InvalidPasswordException extends UserNotFoundException
+{
+
+}