+++ src/Exception/InvalidRegistrationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace HeimrichHannot\LoginRegistrationBundle\Exception;
+
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
+
+class InvalidRegistrationException extends UserNotFoundException
+{
+}