+++ src/Security/User/ContaoUserProviderDecorator.php
@@ -6,13 +6,17 @@ use Contao\Controller;
 use Contao\CoreBundle\Exception\ResponseException;
 use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\CoreBundle\Security\User\ContaoUserProvider;
+use Contao\FormPassword;
 use Contao\FrontendTemplate;
 use Contao\Input;
 use Contao\ModuleModel;
 use Contao\User;
 use HeimrichHannot\LoginRegistrationBundle\Controller\FrontendModule\LoginRegistrationModuleController;
 use HeimrichHannot\LoginRegistrationBundle\Event\AdjustUsernameEvent;
-use HeimrichHannot\LoginRegistrationBundle\Proxy\RegistrationProxy;
+use HeimrichHannot\LoginRegistrationBundle\Exception\InvalidPasswordException;
+use HeimrichHannot\LoginRegistrationBundle\Exception\InvalidRegistrationConfigurationException;
+use HeimrichHannot\LoginRegistrationBundle\Registration\RegistrationProxy;
+use HeimrichHannot\LoginRegistrationBundle\EventListener\Contao\ParseWidgetListener;
 use HeimrichHannot\LoginRegistrationBundle\Security\RegistrationUtils;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
@@ -30,6 +34,7 @@ class ContaoUserProviderDecorator implements UserProviderInterface, PasswordUpgr
         private readonly ContaoFramework $framework,
         private readonly EventDispatcherInterface $eventDispatcher,
         private readonly RegistrationUtils $registrationUtils,
+        private readonly ParseWidgetListener $parseWidgetListener
     ) {
     }
 
@@ -72,7 +77,7 @@ class ContaoUserProviderDecorator implements UserProviderInterface, PasswordUpgr
 
         try {
             $user = $this->contaoUserProvider->loadUserByIdentifier($event->getUsername());
-        } catch (UserNotFoundException|UsernameNotFoundException) {
+        } catch (UserNotFoundException|UsernameNotFoundException $e) {
             return $this->applyDirectRegistration($event->getUsername());
         }
 
@@ -83,21 +88,19 @@ class ContaoUserProviderDecorator implements UserProviderInterface, PasswordUpgr
     {
         $this->framework->initialize();
 
-        $userNotFoundException = new UserNotFoundException(sprintf('Could not find user "%s"', $identifier));
-
         $request = $this->requestStack->getCurrentRequest();
         if (!$request) {
-            throw $userNotFoundException;
+            throw new InvalidRegistrationConfigurationException('No request available.');
         }
 
         $moduleId = str_replace('tl_login_', '', $request->request->get('FORM_SUBMIT', ''));
         if (empty($moduleId) || !is_numeric($moduleId)) {
-            throw $userNotFoundException;
+            throw new InvalidRegistrationConfigurationException('No module id found.');
         }
 
         $moduleModel = ModuleModel::findByPk((int) $moduleId);
         if (!$moduleModel || LoginRegistrationModuleController::TYPE !== $moduleModel->type) {
-            throw $userNotFoundException;
+            throw new InvalidRegistrationConfigurationException('No valid module found.');
         }
 
         if (!isset($GLOBALS['objPage'])) {
@@ -121,11 +124,22 @@ class ContaoUserProviderDecorator implements UserProviderInterface, PasswordUpgr
         $registrationModule->Template = new FrontendTemplate();
 
         try {
+            $this->parseWidgetListener->clear();
             $registrationModule->runCompile();
         } catch (ResponseException) {
         }
 
-        $user = $this->loadUserByIdentifier($identifier);
+        foreach ($this->parseWidgetListener->getWidgets() as $widget) {
+            if ($widget->hasErrors()) {
+                if ($widget instanceof FormPassword) {
+                    throw new InvalidPasswordException($widget->getErrorAsString());
+                }
+
+                throw new UserNotFoundException($widget->getErrorsAsString());
+            }
+        }
+
+        $user = $this->contaoUserProvider->loadUserByIdentifier($identifier);
 
         $this->registrationUtils->setLastRegisteredUser($user);